refactor isFromEpicOrBanner logic
